improved relics

+++ include/cub3D.h
@@ -155,8 +155,8 @@ int			get_blue(int rgba);
 int			get_alpha(int rgba);
 
 //utils/utils.c
-int			min(int a, int b);
-int			max(int a, int b);
+double		min(double a, double b);
+double		max(double a, double b);
 
 //hud.c
 void		setup_hud(t_window *window);
